Move code from cmd/telepresence/main.go to pkg/client/cli/main.go
Just moves all code except a tiny dispatch method. No code-changes.
